requirements.txt
ttkbootstrap
pyttsx3
reportlab
Pillow
pyperclip
transformers
requests
torch
speechrecognition
pyaudio